check if pids.txt exists before trying to kill existing processes, don't let run script fail if kill commands fail

+++ run-all.sh
@@ -7,17 +7,18 @@ LOGS_DIR='./logs'
 
 
 
-echo "killing existing executables if running"
+set +e
+if [ -f $LOGS_DIR/pids.txt]; then
+  echo "killing existing executables if running"
+  OLD_APP_PID=$(awk 'NR == 1 {print $3}' $LOGS_DIR/pids.txt)
+  OLD_SLASH_COMMAND_PID=$(awk 'NR == 2 {print $4}' $LOGS_DIR/pids.txt)
+  kill "$OLD_APP_PID"
+  kill "$OLD_SLASH_COMMAND_PID"
+fi
 
-OLD_APP_PID=$(awk 'NR == 1 {print $3}' $LOGS_DIR/pids.txt)
-OLD_SLASH_COMMAND_PID=$(awk 'NR == 2 {print $4}' $LOGS_DIR/pids.txt)
-kill "$OLD_APP_PID"
-kill "$OLD_SLASH_COMMAND_PID"
+set -e
 
 cd ./scripts
-
-
-
 ./install_deps.sh
 ./build.sh
 
@@ -27,19 +28,10 @@ if [ ! -d "$LOGS_DIR" ]; then
   mkdir "$LOGS_DIR"
 fi
 
-echo "killing existing executables if running"
-OLD_APP_PID=(awk 'NR == 0 {print $2}' $LOGS_DIR/pids.txt)
-echo "$OLD_APP_PID"
-
-set +e
 echo "starting executables"
-
 ./scripts/executables/llamabot --token=$TOKEN --port=8081 >>$LOGS_DIR/app.log &
-
 APP_PID=$!
-
 ./scripts/executables/llamabot-slash-command --token=$TOKEN --port=8082 >>$LOGS_DIR/slash-command.log &
-
 SLASH_COMMAND_PID=$!
 
 echo "executables started, writing PIDs to $LOGS_DIR/pids.txt"
